I'm reverting the various changes to the throughput test programs since as far as I can see, these didn't work any more. At the very least, the command line API was broken and forced the user to enter new, exotic arguments. Patches should not break existing APIs. But also, the internals of these programs had become weird.
If we want to build more complex performance tests, that's fine, but we should
make new programs, not break the old ones. We need minimal, safe performance
tests in 0MQ.

Also, the code was quite horrid. So it's gone. If anyone wants to bring it back
please make the code neat, and build new APIs instead of breaking the old ones.

Cheers
Pieter