tools/: Link curve_keygen with a c++ linker
When trying to linking curve_keygen as if it was a C program, linking
fails (at least with gcc 4.8 on Debian testing) because it can't find
C++ symbols. Linking with g++ instead fixes the issue.

To achieve this, the source is renamed to curve_keygen.cpp, and
tools/Makefile.am is updated accordingly.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>