Check for libtoolize instead of libtool on Linux
autogen.sh has a check for the libtool binary as a mean to
check if libtool is available. But distributions like Debian and
Ubuntu are splitting the libtool package, and the libtool binary is now
in a separate package. What autoconf actually need is not the libtool
binary, but libtoolize and other macro files. So check for libtoolize
instead. On the other hand, OSX only ships libtool, not libtoolize,
and uses a pre-generated libtool script to build. So check for
libtoolize first and then for libtool, and fail if neither can be
found.